refactor: divide code into diff files
