+++ async-race/.eslintrc.json
@@ -42,7 +42,7 @@
         }
       ],
       "max-lines-per-function": ["error", { "max": 40 }],
-      "no-magic-numbers": "error",
+      "no-magic-numbers": ["error", { "ignore": [0, 1] }],
       "@typescript-eslint/explicit-function-return-type": "error",
       "@typescript-eslint/no-unnecessary-type-assertion": "error",
       "@typescript-eslint/no-non-null-assertion": "error",
