+++ async-race/src/types/httpstatuscode.ts
@@ -1,0 +1,9 @@
+const OK = 200;
+const CREATED = 201;
+const NOT_FOUND = 404;
+
+export const enum HTTPStatusCode {
+  _OK = OK,
+  _CREATED = CREATED,
+  _NOT_FOUND = NOT_FOUND,
+}
